+++ spec/resources/numismatics/issues/numismatics/issues_feature_spec.rb
@@ -242,4 +242,95 @@ RSpec.feature "Numismatics::Issues" do
       expect(last_attribute["attached"]).to be false
     end
   end
+
+  describe "form editing", js: true do
+    let(:adapter) { Valkyrie::MetadataAdapter.find(:index_solr) }
+    let(:numismatic_issue) do
+      FactoryBot.create_for_repository(:numismatic_issue,
+                                       color: "green",
+                                       denomination: "1/2 Penny",
+                                       edge: "edge",
+                                       metal: "copper",
+                                       object_type: "coin",
+                                       obverse_figure: "obv figure",
+                                       obverse_orientation: "obv orientation",
+                                       obverse_part: "obv part",
+                                       reverse_figure: "rev figure",
+                                       reverse_orientation: "rev orientation",
+                                       reverse_part: "rev part",
+                                       shape: "round")
+    end
+
+    it "displays select boxes for some properties" do
+      visit new_numismatics_issue_path
+      expect(page).to have_css("#numismatics_issue_color.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_denomination.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_edge.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_metal.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_object_type.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_obverse_figure.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_obverse_orientation.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_obverse_part.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_reverse_figure.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_reverse_orientation.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_reverse_part.select2", visible: false)
+      expect(page).to have_css("#numismatics_issue_shape.select2", visible: false)
+    end
+
+    context "when Issues have been saved" do
+      let(:persisted) do
+        change_set = Numismatics::IssueChangeSet.new(numismatic_issue)
+        change_set_persister.save(change_set: change_set)
+      end
+
+      it "permits users to select from existing object types" do
+        visit edit_numismatics_issue_path(id: persisted.id)
+
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#color", visible: false)
+        expect(hidden["value"]).to eq("green")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#denomination", visible: false)
+        expect(hidden["value"]).to eq("1/2 Penny")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#edge", visible: false)
+        expect(hidden["value"]).to eq("edge")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#metal", visible: false)
+        expect(hidden["value"]).to eq("copper")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#object_type", visible: false)
+        expect(hidden["value"]).to eq("coin")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#obverse_figure", visible: false)
+        expect(hidden["value"]).to eq("obv figure")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#obverse_orientation", visible: false)
+        expect(hidden["value"]).to eq("obv orientation")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#obverse_part", visible: false)
+        expect(hidden["value"]).to eq("obv part")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#reverse_figure", visible: false)
+        expect(hidden["value"]).to eq("rev figure")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#reverse_orientation", visible: false)
+        expect(hidden["value"]).to eq("rev orientation")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#reverse_part", visible: false)
+        expect(hidden["value"]).to eq("rev part")
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#shape", visible: false)
+        expect(hidden["value"]).to eq("round")
+
+        expect(page).to have_selector("option", text: "coin")
+        expect(page).to have_selector("option", text: "1/2 Penny")
+        expect(page).to have_selector("option", text: "copper")
+        expect(page).to have_selector("option", text: "round")
+        expect(page).to have_selector("option", text: "green")
+        expect(page).to have_selector("option", text: "edge")
+        expect(page).to have_selector("option", text: "obv figure")
+        expect(page).to have_selector("option", text: "obv orientation")
+        expect(page).to have_selector("option", text: "obv part")
+        expect(page).to have_selector("option", text: "rev figure")
+        expect(page).to have_selector("option", text: "rev orientation")
+        expect(page).to have_selector("option", text: "rev part")
+      end
+
+      it "persists already saved denominations" do
+        visit edit_numismatics_issue_path(id: persisted.id)
+
+        hidden = page.find("body #main form.edit_numismatics_issue input[type='hidden']#denomination", visible: false)
+        expect(hidden["value"]).to eq("1/2 Penny")
+      end
+    end
+  end
 end